requirements.txt
agentic-framework==0.3.8
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorful==0.5.7
compressed-rtf==1.0.7
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
distlib==0.4.0
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
et-xmlfile==2.0.0
extract-msg==0.29.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imapclient==2.1.0
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
litellm==1.74.7
lxml==6.0.0
markdown-it-py==3.0.0
markdownify==1.2.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
numpy==2.2.5
olefile==0.47
openai==1.75.0
opencensus==0.11.4
opencensus-context==0.1.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer-six==20191110
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
prometheus-client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py-spy==0.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
ray==2.45.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
speechrecognition==3.14.3
sqlalchemy==2.0.41
sqlmodel==0.0.24
sse-starlette==2.3.4
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
textract-supercog==1.6.5.post1
thespian==4.0.1
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
typer==0.15.4
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.32.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
